requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
xgboost
lightgbm
shap
joblib
pyyaml
jupyter
notebook
